# Parallelising fan-data to proj-data function in ML_norm.cxx 