--- a/documentation/release_5.2.htm
+++ b/documentation/release_5.2.htm
@@ -30,6 +30,9 @@
 
 <h3>New functionality</h3>
 <ul>
+  <li>Parallelised function <code>set_fan_data_add_gaps_help</code> across segments to reduce computation time.
+  <br /><a href="https://github.com/UCL/STIR/pull/1168/">PR #1168</a>.
+  </li>
 </ul>
 
 <h3>New examples</h3>