Relates to issue: https://github.com/UCL/STIR/issues/1167