@markus-jehl 2023-02-28
Before the change, the following calls were timed as:
- make_fan_data_remove_gaps: 2.8 s
- apply_efficiencies: 0.7 s
- apply_geo_norm: 2.1 s
- set_fan_data_add_gaps: 3.1 s

After the change, they were much faster:
- make_fan_data_remove_gaps: 1 s
- apply_efficiencies: 0.1 s
- apply_geo_norm: 0.4 s
- set_fan_data_add_gaps: 0.6 s