@KrisThielemans 2023-03-03
not sure if I agree. The lines I quoted are in the 1st loop and update `new_fan_data`. They are problematic. There's a loop [here](https://github.com/UCL/STIR/blob/fda27e0c8018a6fa0a531ddf3d5e2fb6755ed89c/src/buildblock/ML_norm.cxx#L1399) (3rd one?) which updated `work`. The [next one](https://github.com/UCL/STIR/blob/fda27e0c8018a6fa0a531ddf3d5e2fb6755ed89c/src/buildblock/ML_norm.cxx#L1437) updates `fan_data`. No?