@KrisThielemans 2023-03-06
but updating `fan_data` is just as problematic, certainly see as we update it symmetrically (`fan_data(new_ra, new_a, new_rb, new_b) =  fan_data(new_rb, new_b, new_ra, new_a) = ...`). How do we know that another thread isn't access the "symmetric" version, or just the neghbouring one, with therefore the potential for memory corruption?

It might be alright as all the index access in the 4D array is read-only, and we only update the 1D vectors, but how do know that those 1D arrays are not adjacent for different segments?

Of course, it seems pretty unlikely that this would generate a race-condition, but I don't think we have a solid guarantee.

I don't have access to any tools to check thread-safety sadly.