@KrisThielemans 2023-03-06
yeah.. by the way, there's no reason for this code
```c++
 shared_ptr<SegmentBySinogram<float> > segment_ptr;
 segment_ptr.reset(new SegmentBySinogram<float>(proj_data.get_segment_by_sinogram(bin.segment_num())));
```
the following is essentially equivalent but much clearer
``c++
   const auto segment(proj_data.get_segment_by_sinogram(bin.segment_num()));
```
(I'm hoping I didn't write those lines)

Note that this is going to make TSAN problems disappear.