@KrisThielemans 2023-03-13
> I managed to get OpenMP working with TSAN for a very simple dummy for loop just printing something to cout.

well, that's a bit strange. Writing to `cout` is most definitely not thread-safe with `critical` section, so it should have complained! I've tried to find some doc on clang/TSAM/OpenMP but gave up. The only info i could find (but it's old) is that you need to build your own libgomp, and then `LD_PRELOAD` it.

>  "atomic read" on the fan_data doesn't even compile, because it expects this to work on simple operations such as "v=x", not on complex functions. Even on the "[]"-style indexing in VectorWithOffset it complained

I'm not surprised by the `atomic` restrictions for `fan_data`, but am disappointed it cannot handle `VectorWithOffset::operator[]`. Does it work with `std::vector::operator[]` by the way?

We don't need atomic reads anywhere. They are only necessary when someone else can write to that memory. So, in the second loop, we read `fan_data` in parallel (fine), and write to a thread-local segment (fine), and then call `set_segment`, which should have its own internal critical section., so should be fine as well. You can always check by adding your own critical section for the `set_segment`.

Updating `fan_data` though is harder.